requirements.txt
# sudo apt update
# sudo apt install libpq-dev python3-dev

pandas
pyarrow
pyodbc
python-dotenv
sqlalchemy
geoalchemy2
psycopg2
pytest
azure-storage-blob
